Eliminate the pubSubChannel and global emitter
